utils/android/workloads: move workloads loading at module initialisation
This "seems" to be required to make workloads plugins visible from
external python scripts.
